Update introduction.md
- removed bold from h1 and h2 titles which I added last time.
I didn't know about theme using for styling than markdown.
- nested functions & objects done.
I think it should be good for now. Let me know if any topics in introduction.md can be improved/added.
Maybe "control structures" can be splitted into "if's" and "loops", not sure how to, but I'd leave it as it is for now.